+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/DebugGetRawReceipts.java
@@ -14,10 +14,11 @@
  */
 package org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods;
 
+import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.ethereum.api.jsonrpc.RpcMethod;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.JsonRpcRequestContext;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.exception.InvalidJsonRpcParameters;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.BlockParameter;
+import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.BlockParameterOrBlockHash;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.JsonRpcParameter.JsonRpcParameterException;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.RpcErrorType;
 import org.hyperledger.besu.ethereum.api.query.BlockchainQueries;
@@ -28,12 +29,10 @@ import org.hyperledger.besu.ethereum.rlp.RLP;
 
 import java.util.List;
 
-import com.google.common.base.Suppliers;
-
-public class DebugGetRawReceipts extends AbstractBlockParameterMethod {
+public class DebugGetRawReceipts extends AbstractBlockParameterOrBlockHashMethod {
 
   public DebugGetRawReceipts(final BlockchainQueries blockchain) {
-    super(Suppliers.ofInstance(blockchain));
+    super(blockchain);
   }
 
   @Override
@@ -42,21 +41,24 @@ public class DebugGetRawReceipts extends AbstractBlockParameterMethod {
   }
 
   @Override
-  protected BlockParameter blockParameter(final JsonRpcRequestContext request) {
+  protected BlockParameterOrBlockHash blockParameterOrBlockHash(
+      final JsonRpcRequestContext request) {
     try {
-      return request.getRequiredParameter(0, BlockParameter.class);
+      return request.getRequiredParameter(0, BlockParameterOrBlockHash.class);
     } catch (JsonRpcParameterException e) {
       throw new InvalidJsonRpcParameters(
-          "Invalid block parameter (index 0)", RpcErrorType.INVALID_BLOCK_PARAMS, e);
+          "Invalid block or block hash parameter (index 0)", RpcErrorType.INVALID_BLOCK_PARAMS, e);
     }
   }
 
   @Override
-  protected Object resultByBlockNumber(
-      final JsonRpcRequestContext request, final long blockNumber) {
+  protected Object resultByBlockHash(final JsonRpcRequestContext request, final Hash blockHash) {
+    if (Hash.EMPTY.getBytes().equals(blockHash.getBytes())) {
+      return null;
+    }
     return getBlockchainQueries()
-        .getBlockHashByNumber(blockNumber)
-        .flatMap(blockHash -> getBlockchainQueries().getBlockchain().getTxReceipts(blockHash))
+        .getBlockchain()
+        .getTxReceipts(blockHash)
         .map(this::toRLP)
         .orElse(null);
   }